Able to attach parts, affects weapon stats.

+++ Physics/Source/PartScript.cpp
@@ -2,7 +2,7 @@
 
 
 
-PartScript::PartScript(PART_TYPE type, SLOT_TYPE slot, float Multiplier, bool isAttached, float durability) 
+PartScript::PartScript(PART_TYPE type, SLOT_TYPE slot, double Multiplier, bool isAttached, float durability) 
 	:m_partType (type),
 	 m_slotType (slot),
 	 m_fStatMultiplier(Multiplier),
